+++ sketch-core/src/androidTest/kotlin/com/github/panpf/sketch/core/test/decode/internal/OpenGLTextureHelperTest.kt
@@ -15,7 +15,6 @@
  */
 package com.github.panpf.sketch.core.test.decode.internal
 
-import android.os.Build
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.panpf.sketch.decode.internal.OpenGLTextureHelper
 import org.junit.Assert
@@ -27,11 +26,10 @@ class OpenGLTextureHelperTest {
 
     @Test
     fun test() {
-        if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.JELLY_BEAN) {
-            Assert.assertNull(OpenGLTextureHelper.maxSize)
-        } else {
-            val maxSize = OpenGLTextureHelper.maxSize
-            Assert.assertTrue(arrayOf(4096, 16384).any { it == maxSize })
-        }
+        val maxSize = OpenGLTextureHelper.maxSize
+        Assert.assertTrue(
+            /* message = */ "maxSize=$maxSize",
+            /* condition = */ arrayOf(2048, 4096, 8192, 16384).any { it == maxSize }
+        )
     }
 }