+++ sketch-zoom/src/main/kotlin/com/github/panpf/sketch/zoom/internal/TileUtils.kt
@@ -16,7 +16,6 @@
 package com.github.panpf.sketch.zoom.internal
 
 import android.graphics.Rect
-import com.github.panpf.sketch.decode.internal.maxBitmapSize
 import com.github.panpf.sketch.util.Size
 import com.github.panpf.sketch.zoom.Tile
 import kotlin.math.abs
@@ -24,9 +23,8 @@ import kotlin.math.ceil
 
 internal fun initializeTileMap(imageSize: Size, tileMaxSize: Size): Map<Int, List<Tile>> {
     /* The core rules are: The size of each tile does not exceed tileMaxSize */
-    val maxBitmapSize = maxBitmapSize
-    val tileMaxWith = tileMaxSize.width.coerceAtMost(maxBitmapSize.width)
-    val tileMaxHeight = tileMaxSize.height.coerceAtMost(maxBitmapSize.height)
+    val tileMaxWith = tileMaxSize.width
+    val tileMaxHeight = tileMaxSize.height
     val tileMap = HashMap<Int, List<Tile>>()
 
     var sampleSize = 1